remove unused and rename
